Implement all splines with unit tests
Implement unit tests for all the splines, to ensure that the return
values are always in the correct range, and that the log det is a
number.

Additionally, write unit tests for all the coupling layers to ensure
that the forward and inverse transform are consistent, and that the
forward log det is equal to the negative of the inverse log det.

During this testing, bugs were found in almost all the splines. The
cubic spline is finally coded up correctly and runs. Finally, all of the
results are consistent to double precision accuracy.