Improve formatting of the code.
Work on ensuring that all the code is formatted according to PEP8
standards, and that pylint does not throw errors.

NOTE: Most of the files include lines turning off some of the erorrs and
warnings from pylint. This is done to ensure that the code does not
introduce arbitrary names of variables that currently do not conform to
the snake_case standard. Additionally, many of the functions take too
many arguments, working out a way to resolve this would clutter the code
more than needed.