Implement acceptance_calc and loss_func flag
Add in the ability to calculate the acceptance based off of the formulas
that were discussed:

        N = \frac{\eps^-2}{\eta(N)}

where N is the number of weighted events required to obtain a precision
of \eps on the total cross-section (\eps = \delta\sigma / \sigma) for
unweighted events, and \eta(N) is the acceptance efficiency for
generating N weighted events. This then allows us to get an estimate of
what the maximum value of the function should be.

Additionally, in this commit, the ability to send in a flag to determine
the loss function is added. Currently, the options are either kl or
chi2.