Update unit tests
Update unit tests for both the divergences and the integrator classes.
The tests now pass for all possible divergences. This should have a
coverage now of about 95%.

Additionally, add in an example script to use the integrator. This makes
the flow/integration/integrator.py file much cleaner to look at and
maintain.